Move getChangedModules() to the base class
